src/main/java/at/medunigraz/imi/bst/trec/query/TemplateQueryDecorator.java
package at.medunigraz.imi.bst.trec.query;

import java.io.File;
import java.io.IOException;
import java.util.List;

import org.apache.commons.io.FileUtils;

import at.medunigraz.imi.bst.trec.model.Result;
import at.medunigraz.imi.bst.trec.model.Topic;

public class TemplateQueryDecorator extends MapQueryDecorator {
	
	private File template;

	public TemplateQueryDecorator(File template, Query decoratedQuery) {
		super(decoratedQuery);
		this.template = template;
		loadTemplate();
	}

	@Override
	public List<Result> query(Topic topic) {
		loadTemplate();
		map(topic.getAttributes());
		return decoratedQuery.query(topic);
	}

	private String readTemplate(File template) {
		String ret = "";
		try {
			ret = FileUtils.readFileToString(template, "UTF-8");
		} catch (IOException e) {
			e.printStackTrace();
		}
		return ret;
	}
	
	private void loadTemplate() {
		setJSONQuery(readTemplate(template));
	}
	
	@Override
	protected String getMyName() {
		return getSimpleClassName() + "(" + template.getName() + ")";
	}

}
